Added FAQ component. Reorganized folders. Moved data to own file
