+++ client/src/app/App.js
@@ -1,5 +1,5 @@
 import "./App.css";
-import Home from "../pages/home";
+import Home from "../pages/home/home";
 
 function App() {
   return (
