+++ client/src/components/home/signUp.js
@@ -1,6 +1,6 @@
 import React from "react";
 import { Button } from "semantic-ui-react";
-import "../assets/styles.css";
+import "../../assets/styles.css";
 
 const SignUp = () => {
   return (
