--- client/src/components/oneStep.js
@@ -1,9 +1,0 @@
-import React from "react";
-
-const OneStep = (step) => {
-  const { stepNumber, title, subtitle } = step;
-
-  return <div>{stepNumber}</div>;
-};
-
-export default OneStep;
